move hostdb to renter subdir
