+++ modules/renter/hostdb/hostentry.go
