+++ modules/renter/hostdb/hostweight_test.go
